+++ src/client/pages/NotFound.tsx
@@ -1,6 +1,7 @@
-import { Box, Button, Heading, Text, VStack } from '@chakra-ui/react';
+import { Box, Heading, Text, VStack } from '@chakra-ui/react';
 import { Link as RouterLink } from 'react-router';
 import { ROUTES } from '../utilities/constants';
+import { AnimatedButton } from '../ui/components/animated-button';
 import { PageLayout } from '../ui/layout/page-layout';
 import { PageMeta } from '../ui/components/page-meta';
 
@@ -26,12 +27,12 @@ const NotFound = () => {
         </Box>
 
         <VStack gap={3} pt={4}>
-          <Button asChild colorScheme="blue" size="lg">
+          <AnimatedButton asChild colorScheme="blue" size="lg">
             <RouterLink to={ROUTES.HOME}>Go to Home Page</RouterLink>
-          </Button>
-          <Button asChild variant="outline" size="md">
+          </AnimatedButton>
+          <AnimatedButton asChild variant="outline" size="md">
             <RouterLink to={ROUTES.ABOUT}>Learn About This Project</RouterLink>
-          </Button>
+          </AnimatedButton>
         </VStack>
       </VStack>
     </PageLayout>